requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
google-cloud-aiplatform==1.35.0
python-dotenv==1.0.0
httpx==0.25.0
anthropic[vertex]==0.18.1  # Optional: for using the official Anthropic SDK
